FoodTruckFinder.FoodTruck (FoodTruckFinder v0.1.0)

This module is just to model the data in the db and validate any changesets.

If you are looking for business logic, go look at FoodTruckFinder.FoodTrucks module.

Functions

changeset(food_truck, attrs)

Casting and validation of the data for the food_trucks table.

rename_keys(food_trucks)

We need to map json keys to schema keys. This function takes in the result of Jason.decode and remaps it using the keys in the food_trucks schema While we're at it, we will drop all keys we don't care about.

Sample Dataset: %{ "address" => "5 THE EMBARCADERO", "applicant" => "Ziaurehman Amini", "block" => "0234", "blocklot" => "0234017", "cnn" => "30727000", "expirationdate" => "2016-03-15T00:00:00.000", "facilitytype" => "Push Cart", "latitude" => "37.794331003246846", "location" => %{

"human_address" => "{"address": "", "city": "", "state": "", "zip": ""}",
"latitude" => "37.794331003246846",
"longitude" => "-122.39581105302317"

}, "locationdescription" => "MARKET ST: DRUMM ST intersection", "longitude" => "-122.39581105302317", "lot" => "017", "objectid" => "735318", "permit" => "15MFF-0159", "priorpermit" => "0", "received" => "20151231", "schedule" => "http://bsm.sfdpw.org/PermitsTracker/reports/report.aspx?title=schedule&report=rptSchedule&params=permit=15MFF-0159&ExportPDF=1&Filename=15MFF-0159_schedule.pdf", "status" => "REQUESTED", "x" => "6013916.72", "y" => "2117244.027" }